Split up MainActivity.kt
Move MainScreen to its own file and move both into a main package.
