+++ app/src/main/java/com/sats/johnnydeep/main/MainActivity.kt
@@ -1,0 +1,18 @@
+package com.sats.johnnydeep.main
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.core.view.WindowCompat
+
+class MainActivity : ComponentActivity() {
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+
+    WindowCompat.setDecorFitsSystemWindows(window, false)
+
+    setContent {
+      MainScreen()
+    }
+  }
+}
